requirements.txt
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
djangorestframework==3.11.1
cffi==1.14.2
channels==2.4.0
channels-redis==3.1.0
uuid==1.30
psycopg2==2.8.5